pyproject.toml
[project]
name = "encuesta-2026"
version = "0.1.0"
description = "Survey data analysis — AI-delegable tasks in Chilean aquaculture veterinary practice"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "jupyter>=1.1",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[tool.uv]
dev-dependencies = []
